web/lesson14/src/schemas.py
from datetime import date, datetime

from pydantic import BaseModel, EmailStr, Field


class ContactResponse(BaseModel):
    id: int
    first_name: str
    last_name: str
    email: EmailStr
    phone_number: str = Field(default='+380123456789')
    birthday: date


class ContactRequest(BaseModel):
    first_name: str
    last_name: str
    email: EmailStr
    phone_number: str = Field(default='+380123456789')
    birthday: date


class TokenModel(BaseModel):
    access_token: str
    refresh_token: str
    token_type: str = 'bearer'


class UserDb(BaseModel):
    id: int
    username: str
    email: str
    created_at: datetime
    avatar: str

    class Config:
        # orm_mode = True
        from_attributes = True


class UserModel(BaseModel):
    username: str = Field(min_length=5, max_length=16)
    email: str
    password: str = Field(min_length=6, max_length=10)


class UserResponse(BaseModel):
    user: UserDb
    detail: str = 'User successfully created'
