web/lesson14/requirements.txt
fastapi==0.103.2
fastapi-limiter==0.1.5
fastapi-mail==1.4.1
redis==4.6.0
types-redis==4.6.0.7
sqlalchemy==2.0.21
psycopg2-binary==2.9.9
alembic==1.12.0
uvicorn==0.23.2
email_validator==2.0.0.post2
libgravatar==1.0.4
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
pydantic-settings==2.0.3
cloudinary==1.36.0
sphinx==7.2.6
faker==19.9.0
pytest==7.4.2